Feature: test snapshot for form checkout
