+++ practice-one/src/components/common/Card/Card.test.tsx
@@ -4,11 +4,9 @@ import "@testing-library/jest-dom";
 import Card from ".";
 import { VARIANT } from "../../../types/Variant";
 
-const mockedUsedNavigate = jest.fn();
-
 jest.mock("react-router-dom", () => ({
   ...jest.requireActual("react-router-dom"),
-  useNavigate: () => mockedUsedNavigate,
+  useNavigate:jest.fn(),
 }));
 
 describe("Card Component", () => {
